Control the form elements
Add a callback to loading the current poll so that the Poll page can load a local copy of the poll for editing. Add a callback to loading the local copy of the poll to figure out which answers are currently selected. Render the answers with the correct answer already selected if logged in. Add a change handler to allow for changing selection.